move delete agent config and copy config id to new section
